feat(frontend): recreate all components as dumb components
This is a step towards non-logic UI vs state coupled UI and will
hopefully simplify the future development of the frontend. As it
recreates the already existing components in a different namespace there
is a lot of duplication, which will be removed, once the state is
connected to these new components.